refactor mongo schemas and postNewBox
