+++ models/box.js
@@ -1,0 +1,119 @@
+'use strict';
+var mongoose = require('mongoose'),
+  timestamp = require('mongoose-timestamp'),
+  Schema = mongoose.Schema,
+  sensorSchema = require('./sensor').schema,
+  products = require('../products');
+
+//Location schema
+var LocationSchema = new Schema({
+  type: {
+    type: String,
+    required: true,
+    default: 'Feature'
+  },
+  geometry: {
+    type: {
+      type: String,
+      required: true,
+      default: 'Point'
+    },
+    coordinates: {
+      type: Array,
+      required: true
+    }
+  },
+  properties: Schema.Types.Mixed
+});
+LocationSchema.index({ 'geometry': '2dsphere' });
+
+//senseBox schema
+var boxSchema = new Schema({
+  name: {
+    type: String,
+    required: true,
+    trim: true
+  },
+  loc: {
+    type: [LocationSchema],
+    required: true
+  },
+  boxType: {
+    type: String,
+    required: true
+  },
+  exposure: {
+    type: String,
+    required: false
+  },
+  grouptag: {
+    type: String,
+    required: false
+  },
+  model: {
+    type: String,
+    required: false
+  },
+  weblink: {
+    type: String,
+    required: false
+  },
+  description: {
+    type: String,
+    required: false
+  },
+  sensors: [sensorSchema]
+},{ strict: false });
+boxSchema.plugin(timestamp);
+
+boxSchema.add({
+  mqtt: {
+    url: { type: String, trim: true },
+    topic: { type: String, trim: true }
+  }
+});
+
+boxSchema.statics.newFromRequest = function (req) {
+  var boxData = {
+    name: req.params.name,
+    boxType: req.params.boxType,
+    loc: req.params.loc,
+    grouptag: req.params.tag,
+    exposure: req.params.exposure,
+    _id: mongoose.Types.ObjectId(),
+    model: req.params.model,
+    sensors: req.params.sensors
+  };
+
+  // if model is not empty, get sensor definitions from products
+  // otherwise, req.params.sensors should not be empty
+  if (req.params.model) {
+    switch (req.params.model) {
+    case 'homeEthernet':
+      boxData.sensors = products.senseboxhome;
+      break;
+    case 'homeWifi':
+      boxData.sensors = products.senseboxhome;
+      break;
+    case 'basicEthernet':
+      boxData.sensors = products.senseboxbasic;
+      break;
+    default:
+      break;
+    }
+  }
+
+  // assign sensors an mongodb objectid
+  for (var i = boxData.sensors.length - 1; i >= 0; i--) {
+    boxData.sensors[i] = Object.assign(boxData.sensors[i], { _id: mongoose.Types.ObjectId() });
+  }
+
+  return new this(boxData);
+};
+
+var boxModel = mongoose.model('Box', boxSchema);
+
+module.exports = {
+  schema: boxSchema,
+  model: boxModel
+};
