+++ models/sensor.js
@@ -1,0 +1,38 @@
+'use strict';
+var mongoose = require('mongoose'),
+  Schema = mongoose.Schema;
+
+//Sensor schema
+var sensorSchema = new Schema({
+  title: {
+    type: String,
+    required: true,
+    trim: true
+  },
+  unit: {
+    type: String,
+    required: true,
+    trim: true
+  },
+  sensorType: {
+    type: String,
+    required: false,
+    trim: true
+  },
+  icon: {
+    type: String,
+    required: false,
+    trim: true
+  },
+  lastMeasurement: {
+    type: Schema.Types.ObjectId,
+    ref: 'Measurement'
+  }
+});
+
+var sensorModel = mongoose.model('Sensor', sensorSchema);
+
+module.exports = {
+  schema: sensorSchema,
+  model: sensorModel
+};
