+++ models/user.js
@@ -1,0 +1,71 @@
+'use strict';
+var mongoose = require('mongoose'),
+  Schema = mongoose.Schema;
+
+var userSchema = new Schema({
+  firstname: {
+    type: String,
+    required: true,
+    trim: true
+  },
+  lastname: {
+    type: String,
+    required: true,
+    trim: true
+  },
+  email: {
+    type: String,
+    required: true,
+    trim: true
+  },
+  apikey: {
+    type: String,
+    trim: true,
+    minlength: 24,
+    unique: true,
+    required: true
+  },
+  boxes: [
+    {
+      type: String,
+      trim: true
+    }
+  ],
+  language: {
+    type: String,
+    trim: true
+  }
+});
+
+userSchema.statics.createForBoxRequest = function (box, req, cb) {
+  // aparently this.create(<data>) fails with missing key errors in mongoose 4.6.0
+  // for supplied keys
+  return new this({
+    firstname: req.params.user.firstname,
+    lastname: req.params.user.lastname,
+    email: req.params.user.email,
+    apikey: req.params.orderID,
+    boxes: [box._id],
+    language: req.params.user.lang
+  }).save(cb);
+};
+
+var handleE11000 = function (error, res, next) {
+  if (error.name === 'MongoError' && error.code === 11000) {
+    next(new Error('Duplicate senseBox found'));
+  } else {
+    next();
+  }
+};
+
+userSchema.post('save', handleE11000);
+userSchema.post('update', handleE11000);
+userSchema.post('findOneAndUpdate', handleE11000);
+userSchema.post('insertMany', handleE11000);
+
+var userModel = mongoose.model('User', userSchema);
+
+module.exports = {
+  schema: userSchema,
+  model: userModel
+};
